be/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "atc-audio-agent"
version = "1.0.0"
description = "Real-time Air Traffic Control audio processing with AI-powered transcription and language understanding"
authors = [
    {name = "AI Berkeley 2025 Team", email = "team@example.com"}
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "uagents>=0.12.0",
    "groq>=0.28.0",
    "websockets>=12.0",
    "pyaudio>=0.2.14",
    "numpy>=1.24.0",
    "httpx>=0.25.0",
    "aiohttp>=3.8.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=23.0",
    "isort>=5.0",
    "mypy>=1.0",
]

[project.urls]
Homepage = "https://github.com/ai-berkeley-2025/atc-audio-agent"
Repository = "https://github.com/ai-berkeley-2025/atc-audio-agent"
Issues = "https://github.com/ai-berkeley-2025/atc-audio-agent/issues"

[project.scripts]
atc-audio-agent = "scripts.start:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "scripts*"]

[tool.black]
line-length = 88
target-version = ['py39']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true 